requirements.txt
asyncpg==0.30.0
dlib==19.24.6
facenet_pytorch==2.6.0
fastapi==0.115.12
joblib==1.4.2
mediapipe==0.10.21
mtcnn==1.0.0
numpy==2.2.4
pandas==2.2.3
Pillow==11.1.0
pyautogui==0.9.54
pydantic==2.11.3
python-dotenv==1.1.0
python_bcrypt==0.3.2
scikit_learn==1.6.1
streamlit==1.44.1
tensorflow==2.19.0
torch==2.6.0
ultralytics==8.3.105
